+++ src/view/components/IdleCallbackRequestHistory.svelte
@@ -11,7 +11,7 @@
   } from '../../api/settings.ts';
   import { compareByFieldOrder } from '../../api/comparator.ts';
   import { CALLED_ABORTED_TOOLTIP } from '../../api/const.ts';
-  import { delayTitle } from '../../api/time.ts';
+  import { msToHms } from '../../api/time.ts';
   import Variable from './Variable.svelte';
   import Trace from './Trace.svelte';
   import TraceDomain from './TraceDomain.svelte';
@@ -176,7 +176,7 @@
           {/if}
         </td>
         <td class="ta-c"><Variable value={metric.handler} /></td>
-        <td class="ta-r" title={delayTitle(metric.delay)}>{metric.delay}</td>
+        <td class="ta-r" title={msToHms(metric.delay)}>{metric.delay}</td>
         <td class="ta-r">
           {#if metric.online}
             <Variable value={metric.online} />
